verarte-back-end/pyproject.toml
[project]
name = "verarte-back-end"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.9.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "gunicorn>=25.0.1",
    "httpx>=0.28.1",
    "pillow>=12.1.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "resend>=2.21.0",
    "ruff>=0.14.14",
    "tortoise-orm[asyncpg]>=0.25.3",
    "uvicorn[standard]>=0.40.0",
]

[tool.aerich]
tortoise_orm = "app.database_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
